+++ odoo/addons/test_new_api/tests/test_search.py
@@ -815,7 +815,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_city"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -830,9 +830,9 @@ class TestFlushSearch(TransactionCase):
     def test_flush_fields_in_subdomain(self):
         with self.assertQueries(['''
             UPDATE "test_new_api_city"
-            SET "country_id" = "__tmp"."country_id"::int4,
+            SET "country_id" = "__tmp"."country_id"::int8,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "country_id", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -852,7 +852,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_country"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_country"."id" = "__tmp"."id"
         ''', '''
@@ -873,9 +873,9 @@ class TestFlushSearch(TransactionCase):
 
         with self.assertQueries(['''
             UPDATE "test_new_api_city"
-            SET "country_id" = "__tmp"."country_id"::int4,
+            SET "country_id" = "__tmp"."country_id"::int8,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "country_id", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -895,9 +895,9 @@ class TestFlushSearch(TransactionCase):
 
         with self.assertQueries(['''
             UPDATE "test_new_api_payment"
-            SET "move_id" = "__tmp"."move_id"::int4,
+            SET "move_id" = "__tmp"."move_id"::int8,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "move_id", "write_date", "write_uid")
             WHERE "test_new_api_payment"."id" = "__tmp"."id"
         ''', '''
@@ -915,7 +915,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_move"
             SET "tag_repeat" = "__tmp"."tag_repeat"::int4,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "tag_repeat", "write_date", "write_uid")
             WHERE "test_new_api_move"."id" = "__tmp"."id"
         ''', '''
@@ -942,7 +942,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_city"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -959,7 +959,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_city"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -976,7 +976,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_country"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_country"."id" = "__tmp"."id"
         ''', '''
@@ -994,9 +994,9 @@ class TestFlushSearch(TransactionCase):
 
         with self.assertQueries(['''
             UPDATE "test_new_api_city"
-            SET "country_id" = "__tmp"."country_id"::int4,
+            SET "country_id" = "__tmp"."country_id"::int8,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "country_id", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -1027,7 +1027,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_city"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
@@ -1043,7 +1043,7 @@ class TestFlushSearch(TransactionCase):
             UPDATE "test_new_api_city"
             SET "name" = "__tmp"."name"::VARCHAR,
                 "write_date" = "__tmp"."write_date"::timestamp,
-                "write_uid" = "__tmp"."write_uid"::int4
+                "write_uid" = "__tmp"."write_uid"::int8
             FROM (VALUES %s) AS "__tmp"("id", "name", "write_date", "write_uid")
             WHERE "test_new_api_city"."id" = "__tmp"."id"
         ''', '''
